Rename standard.local.ts into locale.ts
* defines a standard interface for locales
