+++ src/assets/locales/en.ts
@@ -1,4 +1,4 @@
-import { LocaleDefinition } from './standard.local'
+import { LocaleDefinition } from './locale'
 
 const en: LocaleDefinition = {
   main: {
