+++ src/assets/locales/locale.ts
