+++ src/assets/locales/ru.ts
@@ -1,4 +1,4 @@
-import { LocaleDefinition } from './standard.local'
+import { LocaleDefinition } from './locale'
 
 // @ts-ignore
 const ru: LocaleDefinition = {}
